64 Valentia Road
"In heaven's name, what for?"

"In search of the body of Lady St. Simon."

Sherlock Holmes leaned back in his chair and laughed heartily.

"Have you dragged the basin of Trafalgar Square fountain?" he asked.

"Why? What do you mean?"

"Because you have just as good a chance of finding this lady in the one as in the other."