3 Wellswood Gardens
"But what other is there?" cried the banker with a gesture of despair. "If his motives were
innocent, why does he not explain them?"

"It is our task to find that out," replied Holmes; "so now, if you please, Mr. Holder, we will set
off for Streatham together, and devote an hour to glancing a little more closely into details."

My friend insisted upon my accompanying them in their expedition, which I was eager enough to do,
for my curiosity and sympathy were deeply stirred by the story to which we had listened. I confess
that the guilt of the banker's son appeared to me to be as obvious as it did to his unhappy father,
but still I had such faith in Holmes' judgment that I felt that there must be some grounds for hope
as long as he was dissatisfied with the accepted explanation. He hardly spoke a word the whole way
out to the southern suburb, but sat with his chin upon his breast and his hat drawn over his eyes,
sunk in the deepest thought. Our client appeared to have taken fresh heart at the little glimpse of
hope which had been presented to him, and he even broke into a desultory chat with me over his
business affairs. A short railway journey and a shorter walk brought us to Fairbank, the modest
residence of the great financier.