98 Moor End
"He brought round both hat and goose to me on Christmas morning, knowing that even the smallest
problems are of interest to me. The goose we retained until this morning, when there were signs
that, in spite of the slight frost, it would be well that it should be eaten without unnecessary
delay. Its finder has carried it off, therefore, to fulfil the ultimate destiny of a goose, while I
continue to retain the hat of the unknown gentleman who lost his Christmas dinner."

"Did he not advertise?"

"No."