73 Harvey Street
"Game for a morning drive?"

"Certainly."

"Then dress. No one is stirring yet, but I know where the stable-boy sleeps, and we shall soon have
the trap out." He chuckled to himself as he spoke, his eyes twinkled, and he seemed a different man
to the sombre thinker of the previous night.

As I dressed I glanced at my watch. It was no wonder that no one was stirring. It was twenty-five
minutes past four. I had hardly finished when Holmes returned with the news that the boy was putting
in the horse.