131 Holmefield Avenue
While Sherlock Holmes had been detailing this singular series of events, we had been whirling
through the outskirts of the great town until the last straggling houses had been left behind, and
we rattled along with a country hedge upon either side of us. Just as he finished, however, we drove
through two scattered villages, where a few lights still glimmered in the windows.

"We are on the outskirts of Lee," said my companion. "We have touched on three English counties in
our short drive, starting in Middlesex, passing over an angle of Surrey, and ending in Kent. See
that light among the trees? That is The Cedars, and beside that lamp sits a woman whose anxious ears
have already, I have little doubt, caught the clink of our horse's feet."

"But why are you not conducting the case from Baker Street?" I asked.