66 Wellswood Gardens
"'Then, good-bye, Mr. Jabez Wilson, and let me congratulate you once more on the important position
which you have been fortunate enough to gain.' He bowed me out of the room and I went home with my
assistant, hardly knowing what to say or do, I was so pleased at my own good fortune.

"Well, I thought over the matter all day, and by evening I was in low spirits again; for I had quite
persuaded myself that the whole affair must be some great hoax or fraud, though what its object
might be I could not imagine. It seemed altogether past belief that anyone could make such a will,
or that they would pay such a sum for doing anything so simple as copying out the 'Encyclopaedia
Britannica.' Vincent Spaulding did what he could to cheer me up, but by bedtime I had reasoned
myself out of the whole thing. However, in the morning I determined to have a look at it anyhow, so
I bought a penny bottle of ink, and with a quill-pen, and seven sheets of foolscap paper, I started
off for Pope's Court.

"Well, to my surprise and delight, everything was as right as possible. The table was set out ready
for me, and Mr. Duncan Ross was there to see that I got fairly to work. He started me off upon the
letter A, and then he left me; but he would drop in from time to time to see that all was right with
me. At two o'clock he bade me good-day, complimented me upon the amount that I had written, and
locked the door of the office after me.

"This went on day after day, Mr. Holmes, and on Saturday the manager came in and planked down four
golden sovereigns for my week's work. It was the same next week, and the same the week after. Every
morning I was there at ten, and every afternoon I left at two. By degrees Mr. Duncan Ross took to
coming in only once of a morning, and then, after a time, he did not come in at all. Still, of
course, I never dared to leave the room for an instant, for I was not sure when he might come, and
the billet was such a good one, and suited me so well, that I would not risk the loss of it.