103 Wantage Close
"But how about Mr. Hosmer Angel? Did he make no attempt to see you?"

"Well, father was going off to France again in a week, and Hosmer wrote and said that it would be
safer and better not to see each other until he had gone. We could write in the meantime, and he
used to write every day. I took the letters in in the morning, so there was no need for father to
know."

"Were you engaged to the gentleman at this time?"

"Oh, yes, Mr. Holmes. We were engaged after the first walk that we took. Hosmer--Mr. Angel--was a
cashier in an office in Leadenhall Street--and--"

"What office?"

"That's the worst of it, Mr. Holmes, I don't know."