27 Sunningdale Drive
"'Tell me all about it,' said I.

"'Well,' said he, showing me the advertisement, 'you can see for yourself that the League has a
vacancy, and there is the address where you should apply for particulars. As far as I can make out,
the League was founded by an American millionaire, Ezekiah Hopkins, who was very peculiar in his
ways. He was himself red-headed, and he had a great sympathy for all red-headed men; so when he died
it was found that he had left his enormous fortune in the hands of trustees, with instructions to
apply the interest to the providing of easy berths to men whose hair is of that colour. From all I
hear it is splendid pay and very little to do.'