63 Ventnor Terrace
"He looked surprised and, as it seemed to me, a little startled at my remark.

"'Photography is one of my hobbies,' said he. 'I have made my dark room up there. But, dear me! what
an observant young lady we have come upon. Who would have believed it? Who would have ever believed
it?' He spoke in a jesting tone, but there was no jest in his eyes as he looked at me. I read
suspicion there and annoyance, but no jest.

"Well, Mr. Holmes, from the moment that I understood that there was something about that suite of
rooms which I was not to know, I was all on fire to go over them. It was not mere curiosity, though
I have my share of that. It was more a feeling of duty--a feeling that some good might come from my
penetrating to this place. They talk of woman's instinct; perhaps it was woman's instinct which gave
me that feeling. At any rate, it was there, and I was keenly on the lookout for any chance to pass
the forbidden door.

"It was only yesterday that the chance came. I may tell you that, besides Mr. Rucastle, both Toller
and his wife find something to do in these deserted rooms, and I once saw him carrying a large black
linen bag with him through the door. Recently he has been drinking hard, and yesterday evening he
was very drunk; and when I came upstairs there was the key in the door. I have no doubt at all that
he had left it there. Mr. and Mrs. Rucastle were both downstairs, and the child was with them, so
that I had an admirable opportunity. I turned the key gently in the lock, opened the door, and
slipped through.