70 Tamworth Drive
"Have you managed it?" asked Holmes.

A loud thudding noise came from somewhere downstairs. "That is Mrs. Toller in the cellar," said she.
"Her husband lies snoring on the kitchen rug. Here are his keys, which are the duplicates of Mr.
Rucastle's."

"You have done well indeed!" cried Holmes with enthusiasm. "Now lead the way, and we shall soon see
the end of this black business."

We passed up the stair, unlocked the door, followed on down a passage, and found ourselves in front
of the barricade which Miss Hunter had described. Holmes cut the cord and removed the transverse
bar. Then he tried the various keys in the lock, but without success. No sound came from within, and
at the silence Holmes' face clouded over.