23 Down Close
I placed my revolver, cocked, upon the top of the wooden case behind which I crouched. Holmes shot
the slide across the front of his lantern and left us in pitch darkness--such an absolute darkness
as I have never before experienced. The smell of hot metal remained to assure us that the light was
still there, ready to flash out at a moment's notice. To me, with my nerves worked up to a pitch of
expectancy, there was something depressing and subduing in the sudden gloom, and in the cold dank
air of the vault.

"They have but one retreat," whispered Holmes. "That is back through the house into Saxe-Coburg
Square. I hope that you have done what I asked you, Jones?"

"I have an inspector and two officers waiting at the front door."