102 Holmefield Avenue
"Why, it is after nine now," I answered. "I should not be surprised if that were he. I thought I
heard a ring."

It was, indeed, our friend the financier. I was shocked by the change which had come over him, for
his face which was naturally of a broad and massive mould, was now pinched and fallen in, while his
hair seemed to me at least a shade whiter. He entered with a weariness and lethargy which was even
more painful than his violence of the morning before, and he dropped heavily into the armchair which
I pushed forward for him.