90 Larch Walk
"Have you heard anything of the case?" he asked.

"Not a word. I have not seen a paper for some days."

"The London press has not had very full accounts. I have just been looking through all the recent
papers in order to master the particulars. It seems, from what I gather, to be one of those simple
cases which are so extremely difficult."

"That sounds a little paradoxical."

"But it is profoundly true. Singularity is almost invariably a clue. The more featureless and
commonplace a crime is, the more difficult it is to bring it home. In this case, however, they have
established a very serious case against the son of the murdered man."

"It is a murder, then?"