88 Balcombe Close
"And the cigar-holder?"

"I could see that the end had not been in his mouth. Therefore he used a holder. The tip had been
cut off, not bitten off, but the cut was not a clean one, so I deduced a blunt pen-knife."

"Holmes," I said, "you have drawn a net round this man from which he cannot escape, and you have
saved an innocent human life as truly as if you had cut the cord which was hanging him. I see the
direction in which all this points. The culprit is--"

"Mr. John Turner," cried the hotel waiter, opening the door of our sitting-room, and ushering in a
visitor.

The man who entered was a strange and impressive figure. His slow, limping step and bowed shoulders
gave the appearance of decrepitude, and yet his hard, deep-lined, craggy features, and his enormous
limbs showed that he was possessed of unusual strength of body and of character. His tangled beard,
grizzled hair, and outstanding, drooping eyebrows combined to give an air of dignity and power to
his appearance, but his face was of an ashen white, while his lips and the corners of his nostrils
were tinged with a shade of blue. It was clear to me at a glance that he was in the grip of some
deadly and chronic disease.