50 Bowler Street
"None save my partner with his family and an occasional friend of Arthur's. Sir George Burnwell has
been several times lately. No one else, I think."

"Do you go out much in society?"

"Arthur does. Mary and I stay at home. We neither of us care for it."

"That is unusual in a young girl."

"She is of a quiet nature. Besides, she is not so very young. She is four-and-twenty."