168 Beaconside
"Yes, it would be as well."

"And I have some business to do which will take some hours. This business at Coburg Square is
serious."

"Why serious?"

"A considerable crime is in contemplation. I have every reason to believe that we shall be in time
to stop it. But to-day being Saturday rather complicates matters. I shall want your help to-night."

"At what time?"