85 Larch Walk
The man sat huddled up in his chair, with his head sunk upon his breast, like one who is utterly
crushed. Holmes stuck his feet up on the corner of the mantelpiece and, leaning back with his hands
in his pockets, began talking, rather to himself, as it seemed, than to us.

"The man married a woman very much older than himself for her money," said he, "and he enjoyed the
use of the money of the daughter as long as she lived with them. It was a considerable sum, for
people in their position, and the loss of it would have made a serious difference. It was worth an
effort to preserve it. The daughter was of a good, amiable disposition, but affectionate and warm-
hearted in her ways, so that it was evident that with her fair personal advantages, and her little
income, she would not be allowed to remain single long. Now her marriage would mean, of course, the
loss of a hundred a year, so what does her stepfather do to prevent it? He takes the obvious course
of keeping her at home and forbidding her to seek the company of people of her own age. But soon he
found that that would not answer forever. She became restive, insisted upon her rights, and finally
announced her positive intention of going to a certain ball. What does her clever stepfather do
then? He conceives an idea more creditable to his head than to his heart. With the connivance and
assistance of his wife he disguised himself, covered those keen eyes with tinted glasses, masked the
face with a moustache and a pair of bushy whiskers, sunk that clear voice into an insinuating
whisper, and doubly secure on account of the girl's short sight, he appears as Mr. Hosmer Angel, and
keeps off other lovers by making love himself."