66 Corbet Close
"I was driven over by my employer, who was as amiable as ever, and was introduced by him that
evening to his wife and the child. There was no truth, Mr. Holmes, in the conjecture which seemed to
us to be probable in your rooms at Baker Street. Mrs. Rucastle is not mad. I found her to be a
silent, pale-faced woman, much younger than her husband, not more than thirty, I should think, while
he can hardly be less than forty-five. From their conversation I have gathered that they have been
married about seven years, that he was a widower, and that his only child by the first wife was the
daughter who has gone to Philadelphia. Mr. Rucastle told me in private that the reason why she had
left them was that she had an unreasoning aversion to her stepmother. As the daughter could not have
been less than twenty, I can quite imagine that her position must have been uncomfortable with her
father's young wife.

"Mrs. Rucastle seemed to me to be colourless in mind as well as in feature. She impressed me neither
favourably nor the reverse. She was a nonentity. It was easy to see that she was passionately
devoted both to her husband and to her little son. Her light grey eyes wandered continually from one
to the other, noting every little want and forestalling it if possible. He was kind to her also in
his bluff, boisterous fashion, and on the whole they seemed to be a happy couple. And yet she had
some secret sorrow, this woman. She would often be lost in deep thought, with the saddest look upon
her face. More than once I have surprised her in tears. I have thought sometimes that it was the
disposition of her child which weighed upon her mind, for I have never met so utterly spoiled and so
ill-natured a little creature. He is small for his age, with a head which is quite
disproportionately large. His whole life appears to be spent in an alternation between savage fits
of passion and gloomy intervals of sulking. Giving pain to any creature weaker than himself seems to
be his one idea of amusement, and he shows quite remarkable talent in planning the capture of mice,
little birds, and insects. But I would rather not talk about the creature, Mr. Holmes, and, indeed,
he has little to do with my story."

"I am glad of all details," remarked my friend, "whether they seem to you to be relevant or not."