65 Glan Road
"You are very kind, Mr. Holmes, but I cannot do that. I shall be true to Hosmer. He shall find me
ready when he comes back."

For all the preposterous hat and the vacuous face, there was something noble in the simple faith of
our visitor which compelled our respect. She laid her little bundle of papers upon the table and
went her way, with a promise to come again whenever she might be summoned.

Sherlock Holmes sat silent for a few minutes with his fingertips still pressed together, his legs
stretched out in front of him, and his gaze directed upward to the ceiling. Then he took down from
the rack the old and oily clay pipe, which was to him as a counsellor, and, having lit it, he leaned
back in his chair, with the thick blue cloud-wreaths spinning up from him, and a look of infinite
languor in his face.

"Quite an interesting study, that maiden," he observed. "I found her more interesting than her
little problem, which, by the way, is rather a trite one. You will find parallel cases, if you
consult my index, in Andover in '77, and there was something of the sort at The Hague last year. Old
as is the idea, however, there were one or two details which were new to me. But the maiden herself
was most instructive."

"You appeared to read a good deal upon her which was quite invisible to me," I remarked.

"Not invisible but unnoticed, Watson. You did not know where to look, and so you missed all that was
important. I can never bring you to realise the importance of sleeves, the suggestiveness of thumb-
nails, or the great issues that may hang from a boot-lace. Now, what did you gather from that
woman's appearance? Describe it."