131 Beaconside
"Well, at least," said Holmes, "you have had your revenge upon them. There can be no question that
it was your oil-lamp which, when it was crushed in the press, set fire to the wooden walls, though
no doubt they were too excited in the chase after you to observe it at the time. Now keep your eyes
open in this crowd for your friends of last night, though I very much fear that they are a good
hundred miles off by now."

And Holmes' fears came to be realised, for from that day to this no word has ever been heard either
of the beautiful woman, the sinister German, or the morose Englishman. Early that morning a peasant
had met a cart containing several people and some very bulky boxes driving rapidly in the direction
of Reading, but there all traces of the fugitives disappeared, and even Holmes' ingenuity failed
ever to discover the least clue as to their whereabouts.

The firemen had been much perturbed at the strange arrangements which they had found within, and
still more so by discovering a newly severed human thumb upon a window-sill of the second floor.
About sunset, however, their efforts were at last successful, and they subdued the flames, but not
before the roof had fallen in, and the whole place been reduced to such absolute ruin that, save
some twisted cylinders and iron piping, not a trace remained of the machinery which had cost our
unfortunate acquaintance so dearly. Large masses of nickel and of tin were discovered stored in an
out-house, but no coins were to be found, which may have explained the presence of those bulky boxes
which have been already referred to.