16 Ventnor Terrace
"But I am so sure that he is innocent. You know what woman's instincts are. I know that he has done
no harm and that you will be sorry for having acted so harshly."

"Why is he silent, then, if he is innocent?"

"Who knows? Perhaps because he was so angry that you should suspect him."

"How could I help suspecting him, when I actually saw him with the coronet in his hand?"

"Oh, but he had only picked it up to look at it. Oh, do, do take my word for it that he is innocent.
Let the matter drop and say no more. It is so dreadful to think of our dear Arthur in a prison!"