96 Harvey Street
"Mr. Hosmer Angel came to the house again and proposed that we should marry before father came back.
He was in dreadful earnest and made me swear, with my hands on the Testament, that whatever happened
I would always be true to him. Mother said he was quite right to make me swear, and that it was a
sign of his passion. Mother was all in his favour from the first and was even fonder of him than I
was. Then, when they talked of marrying within the week, I began to ask about father; but they both
said never to mind about father, but just to tell him afterwards, and mother said she would make it
all right with him. I didn't quite like that, Mr. Holmes. It seemed funny that I should ask his
leave, as he was only a few years older than me; but I didn't want to do anything on the sly, so I
wrote to father at Bordeaux, where the company has its French offices, but the letter came back to
me on the very morning of the wedding."

"It missed him, then?"

"Yes, sir; for he had started to England just before it arrived."