62 Corbet Close
"Quite so. And she was afterwards seen walking into Hyde Park in company with Flora Millar, a woman
who is now in custody, and who had already made a disturbance at Mr. Doran's house that morning."

"Ah, yes. I should like a few particulars as to this young lady, and your relations to her."