57 Wellswood Gardens
"God knows I have!--a trouble which is enough to unseat my reason, so sudden and so terrible is it.
Public disgrace I might have faced, although I am a man whose character has never yet borne a stain.
Private affliction also is the lot of every man; but the two coming together, and in so frightful a
form, have been enough to shake my very soul. Besides, it is not I alone. The very noblest in the
land may suffer unless some way be found out of this horrible affair."

"Pray compose yourself, sir," said Holmes, "and let me have a clear account of who you are and what
it is that has befallen you."

"My name," answered our visitor, "is probably familiar to your ears. I am Alexander Holder, of the
banking firm of Holder & Stevenson, of Threadneedle Street."