128 Valentia Road
"Was your sister dressed?"

"No, she was in her night-dress. In her right hand was found the charred stump of a match, and in
her left a match-box."

"Showing that she had struck a light and looked about her when the alarm took place. That is
important. And what conclusions did the coroner come to?"