31 Corbet Close
"I cannot think."

"When you combine the ideas of whistles at night, the presence of a band of gipsies who are on
intimate terms with this old doctor, the fact that we have every reason to believe that the doctor
has an interest in preventing his stepdaughter's marriage, the dying allusion to a band, and,
finally, the fact that Miss Helen Stoner heard a metallic clang, which might have been caused by one
of those metal bars that secured the shutters falling back into its place, I think that there is
good ground to think that the mystery may be cleared along those lines."

"But what, then, did the gipsies do?"

"I cannot imagine."

"I see many objections to any such theory."

"And so do I. It is precisely for that reason that we are going to Stoke Moran this day. I want to
see whether the objections are fatal, or if they may be explained away. But what in the name of the
devil!"