120 Larch Walk
"And what did you learn from him?"

"Nothing."

"Could he throw no light?"

"None at all. I was inclined to think at one time that he knew who had done it and was screening him
or her, but I am convinced now that he is as puzzled as everyone else. He is not a very quick-witted
youth, though comely to look at and, I should think, sound at heart."

"I cannot admire his taste," I remarked, "if it is indeed a fact that he was averse to a marriage
with so charming a young lady as this Miss Turner."