80 Bowler Street
"Well, to my surprise and delight, everything was as right as possible. The table was set out ready
for me, and Mr. Duncan Ross was there to see that I got fairly to work. He started me off upon the
letter A, and then he left me; but he would drop in from time to time to see that all was right with
me. At two o'clock he bade me good-day, complimented me upon the amount that I had written, and
locked the door of the office after me.

"This went on day after day, Mr. Holmes, and on Saturday the manager came in and planked down four
golden sovereigns for my week's work. It was the same next week, and the same the week after. Every
morning I was there at ten, and every afternoon I left at two. By degrees Mr. Duncan Ross took to
coming in only once of a morning, and then, after a time, he did not come in at all. Still, of
course, I never dared to leave the room for an instant, for I was not sure when he might come, and
the billet was such a good one, and suited me so well, that I would not risk the loss of it.

"Eight weeks passed away like this, and I had written about Abbots and Archery and Armour and
Architecture and Attica, and hoped with diligence that I might get on to the B's before very long.
It cost me something in foolscap, and I had pretty nearly filled a shelf with my writings. And then
suddenly the whole business came to an end."

"To an end?"

"Yes, sir. And no later than this morning. I went to my work as usual at ten o'clock, but the door
was shut and locked, with a little square of cardboard hammered on to the middle of the panel with a
tack. Here it is, and you can read for yourself."