43 Balcombe Close
Mr. Windibank gave a violent start and dropped his gloves. "I am delighted to hear it," he said.

"It is a curious thing," remarked Holmes, "that a typewriter has really quite as much individuality
as a man's handwriting. Unless they are quite new, no two of them write exactly alike. Some letters
get more worn than others, and some wear only on one side. Now, you remark in this note of yours,
Mr. Windibank, that in every case there is some little slurring over of the 'e,' and a slight defect
in the tail of the 'r.' There are fourteen other characteristics, but those are the more obvious."