173 Beaconside
The nobleman swung his glasses a little faster and stared down into the fire. "You see, Mr. Holmes,"
said he, "my wife was twenty before her father became a rich man. During that time she ran free in a
mining camp and wandered through woods or mountains, so that her education has come from Nature
rather than from the schoolmaster. She is what we call in England a tomboy, with a strong nature,
wild and free, unfettered by any sort of traditions. She is impetuous--volcanic, I was about to say.
She is swift in making up her mind and fearless in carrying out her resolutions. On the other hand,
I would not have given her the name which I have the honour to bear"--he gave a little stately
cough--"had not I thought her to be at bottom a noble woman. I believe that she is capable of heroic
self-sacrifice and that anything dishonourable would be repugnant to her."

"Have you her photograph?"

"I brought this with me." He opened a locket and showed us the full face of a very lovely woman. It
was not a photograph but an ivory miniature, and the artist had brought out the full effect of the
lustrous black hair, the large dark eyes, and the exquisite mouth. Holmes gazed long and earnestly
at it. Then he closed the locket and handed it back to Lord St. Simon.