93 Moor End
"Ah, and what did you gather from this allusion to a band--a speckled band?"

"Sometimes I have thought that it was merely the wild talk of delirium, sometimes that it may have
referred to some band of people, perhaps to these very gipsies in the plantation. I do not know
whether the spotted handkerchiefs which so many of them wear over their heads might have suggested
the strange adjective which she used."

Holmes shook his head like a man who is far from being satisfied.