126 Beaconside
"Circumstantial evidence is a very tricky thing," answered Holmes thoughtfully. "It may seem to
point very straight to one thing, but if you shift your own point of view a little, you may find it
pointing in an equally uncompromising manner to something entirely different. It must be confessed,
however, that the case looks exceedingly grave against the young man, and it is very possible that
he is indeed the culprit. There are several people in the neighbourhood, however, and among them
Miss Turner, the daughter of the neighbouring landowner, who believe in his innocence, and who have
retained Lestrade, whom you may recollect in connection with the Study in Scarlet, to work out the
case in his interest. Lestrade, being rather puzzled, has referred the case to me, and hence it is
that two middle-aged gentlemen are flying westward at fifty miles an hour instead of quietly
digesting their breakfasts at home."

"I am afraid," said I, "that the facts are so obvious that you will find little credit to be gained
out of this case."

"There is nothing more deceptive than an obvious fact," he answered, laughing. "Besides, we may
chance to hit upon some other obvious facts which may have been by no means obvious to Mr. Lestrade.
You know me too well to think that I am boasting when I say that I shall either confirm or destroy
his theory by means which he is quite incapable of employing, or even of understanding. To take the
first example to hand, I very clearly perceive that in your bedroom the window is upon the right-
hand side, and yet I question whether Mr. Lestrade would have noted even so self-evident a thing as
that."