11 Cliff Place
"Oh, you mean the little problem of the Grosvenor Square furniture van. That is quite cleared up now
--though, indeed, it was obvious from the first. Pray give me the results of your newspaper
selections."

"Here is the first notice which I can find. It is in the personal column of the Morning Post, and
dates, as you see, some weeks back:

"'A marriage has been arranged [it says] and will, if rumour is correct, very shortly take place,
between Lord Robert St. Simon, second son of the Duke of Balmoral, and Miss Hatty Doran, the only
daughter of Aloysius Doran. Esq., of San Francisco, Cal., U.S.A.'

That is all."

"Terse and to the point," remarked Holmes, stretching his long, thin legs towards the fire.