172 Beaconside
"Here is an advertisement which will interest you," said he. "It appeared in all the papers about a
year ago. Listen to this:

"'Lost, on the 9th inst., Mr. Jeremiah Hayling, aged twenty-six, a hydraulic engineer. Left his
lodgings at ten o'clock at night, and has not been heard of since. Was dressed in--'