89 Glan Road
"At some evidence implicating Flora Millar in the disappearance."

"I am afraid that you will find it difficult."

"Are you, indeed, now?" cried Lestrade with some bitterness. "I am afraid, Holmes, that you are not
very practical with your deductions and your inferences. You have made two blunders in as many
minutes. This dress does implicate Miss Flora Millar."

"And how?"

"In the dress is a pocket. In the pocket is a card-case. In the card-case is a note. And here is the
very note." He slapped it down upon the table in front of him. "Listen to this: