3 Ventnor Terrace
"John Clay, the murderer, thief, smasher, and forger. He's a young man, Mr. Merryweather, but he is
at the head of his profession, and I would rather have my bracelets on him than on any criminal in
London. He's a remarkable man, is young John Clay. His grandfather was a royal duke, and he himself
has been to Eton and Oxford. His brain is as cunning as his fingers, and though we meet signs of him
at every turn, we never know where to find the man himself. He'll crack a crib in Scotland one week,
and be raising money to build an orphanage in Cornwall the next. I've been on his track for years
and have never set eyes on him yet."

"I hope that I may have the pleasure of introducing you to-night. I've had one or two little turns
also with Mr. John Clay, and I agree with you that he is at the head of his profession. It is past
ten, however, and quite time that we started. If you two will take the first hansom, Watson and I
will follow in the second."

Sherlock Holmes was not very communicative during the long drive and lay back in the cab humming the
tunes which he had heard in the afternoon. We rattled through an endless labyrinth of gas-lit
streets until we emerged into Farrington Street.