65 Bowler Street
"They always fill me with a certain horror. It is my belief, Watson, founded upon my experience,
that the lowest and vilest alleys in London do not present a more dreadful record of sin than does
the smiling and beautiful countryside."

"You horrify me!"

"But the reason is very obvious. The pressure of public opinion can do in the town what the law
cannot accomplish. There is no lane so vile that the scream of a tortured child, or the thud of a
drunkard's blow, does not beget sympathy and indignation among the neighbours, and then the whole
machinery of justice is ever so close that a word of complaint can set it going, and there is but a
step between the crime and the dock. But look at these lonely houses, each in its own fields, filled
for the most part with poor ignorant folk who know little of the law. Think of the deeds of hellish
cruelty, the hidden wickedness which may go on, year in, year out, in such places, and none the
wiser. Had this lady who appeals to us for help gone to live in Winchester, I should never have had
a fear for her. It is the five miles of country which makes the danger. Still, it is clear that she
is not personally threatened."

"No. If she can come to Winchester to meet us she can get away."

"Quite so. She has her freedom."