80 Valentia Road
There was the case, and of course there was but one way out of it. Might I not escort her to this
place? And then, as a second thought, why should she come at all? I was Isa Whitney's medical
adviser, and as such I had influence over him. I could manage it better if I were alone. I promised
her on my word that I would send him home in a cab within two hours if he were indeed at the address
which she had given me. And so in ten minutes I had left my armchair and cheery sitting-room behind
me, and was speeding eastward in a hansom on a strange errand, as it seemed to me at the time,
though the future only could show how strange it was to be.

But there was no great difficulty in the first stage of my adventure. Upper Swandam Lane is a vile
alley lurking behind the high wharves which line the north side of the river to the east of London
Bridge. Between a slop-shop and a gin-shop, approached by a steep flight of steps leading down to a
black gap like the mouth of a cave, I found the den of which I was in search. Ordering my cab to
wait, I passed down the steps, worn hollow in the centre by the ceaseless tread of drunken feet; and
by the light of a flickering oil-lamp above the door I found the latch and made my way into a long,
low room, thick and heavy with the brown opium smoke, and terraced with wooden berths, like the
forecastle of an emigrant ship.

Through the gloom one could dimly catch a glimpse of bodies lying in strange fantastic poses, bowed
shoulders, bent knees, heads thrown back, and chins pointing upward, with here and there a dark,
lack-lustre eye turned upon the newcomer. Out of the black shadows there glimmered little red
circles of light, now bright, now faint, as the burning poison waxed or waned in the bowls of the
metal pipes. The most lay silent, but some muttered to themselves, and others talked together in a
strange, low, monotonous voice, their conversation coming in gushes, and then suddenly tailing off
into silence, each mumbling out his own thoughts and paying little heed to the words of his
neighbour. At the farther end was a small brazier of burning charcoal, beside which on a three-
legged wooden stool there sat a tall, thin old man, with his jaw resting upon his two fists, and his
elbows upon his knees, staring into the fire.

As I entered, a sallow Malay attendant had hurried up with a pipe for me and a supply of the drug,
beckoning me to an empty berth.

"Thank you. I have not come to stay," said I. "There is a friend of mine here, Mr. Isa Whitney, and
I wish to speak with him."