141 Holmefield Avenue
"I trust that we are not too late," said he. "I think, Miss Hunter, that we had better go in without
you. Now, Watson, put your shoulder to it, and we shall see whether we cannot make our way in."

It was an old rickety door and gave at once before our united strength. Together we rushed into the
room. It was empty. There was no furniture save a little pallet bed, a small table, and a basketful
of linen. The skylight above was open, and the prisoner gone.

"There has been some villainy here," said Holmes; "this beauty has guessed Miss Hunter's intentions
and has carried his victim off."