5 Valentia Road
"Have you heard anything of the case?" he asked.

"Not a word. I have not seen a paper for some days."

"The London press has not had very full accounts. I have just been looking through all the recent
papers in order to master the particulars. It seems, from what I gather, to be one of those simple
cases which are so extremely difficult."