177 Larch Walk
"That will await him when he enters port," said he, chuckling. "It may give him a sleepless night.
He will find it as sure a precursor of his fate as Openshaw did before him."

"And who is this Captain Calhoun?"

"The leader of the gang. I shall have the others, but he first."

"How did you trace it, then?"