8 Wilson Gardens
"No, alone. She was very independent in little matters like that. Then, after we had sat down for
ten minutes or so, she rose hurriedly, muttered some words of apology, and left the room. She never
came back."

"But this maid, Alice, as I understand, deposes that she went to her room, covered her bride's dress
with a long ulster, put on a bonnet, and went out."

"Quite so. And she was afterwards seen walking into Hyde Park in company with Flora Millar, a woman
who is now in custody, and who had already made a disturbance at Mr. Doran's house that morning."