54 Wilson Gardens
"No; a client. It seems that a young lady has arrived in a considerable state of excitement, who
insists upon seeing me. She is waiting now in the sitting-room. Now, when young ladies wander about
the metropolis at this hour of the morning, and knock sleepy people up out of their beds, I presume
that it is something very pressing which they have to communicate. Should it prove to be an
interesting case, you would, I am sure, wish to follow it from the outset. I thought, at any rate,
that I should call you and give you the chance."

"My dear fellow, I would not miss it for anything."

I had no keener pleasure than in following Holmes in his professional investigations, and in
admiring the rapid deductions, as swift as intuitions, and yet always founded on a logical basis
with which he unravelled the problems which were submitted to him. I rapidly threw on my clothes and
was ready in a few minutes to accompany my friend down to the sitting-room. A lady dressed in black
and heavily veiled, who had been sitting in the window, rose as we entered.

"Good-morning, madam," said Holmes cheerily. "My name is Sherlock Holmes. This is my intimate friend
and associate, Dr. Watson, before whom you can speak as freely as before myself. Ha! I am glad to
see that Mrs. Hudson has had the good sense to light the fire. Pray draw up to it, and I shall order
you a cup of hot coffee, for I observe that you are shivering."