122 Wilson Gardens
"My dear fellow, I would not miss it for anything."

I had no keener pleasure than in following Holmes in his professional investigations, and in
admiring the rapid deductions, as swift as intuitions, and yet always founded on a logical basis
with which he unravelled the problems which were submitted to him. I rapidly threw on my clothes and
was ready in a few minutes to accompany my friend down to the sitting-room. A lady dressed in black
and heavily veiled, who had been sitting in the window, rose as we entered.

"Good-morning, madam," said Holmes cheerily. "My name is Sherlock Holmes. This is my intimate friend
and associate, Dr. Watson, before whom you can speak as freely as before myself. Ha! I am glad to
see that Mrs. Hudson has had the good sense to light the fire. Pray draw up to it, and I shall order
you a cup of hot coffee, for I observe that you are shivering."