43 Glan Road
A small side door led into the whitewashed corridor from which the three bedrooms opened. Holmes
refused to examine the third chamber, so we passed at once to the second, that in which Miss Stoner
was now sleeping, and in which her sister had met with her fate. It was a homely little room, with a
low ceiling and a gaping fireplace, after the fashion of old country-houses. A brown chest of
drawers stood in one corner, a narrow white-counterpaned bed in another, and a dressing-table on the
left-hand side of the window. These articles, with two small wicker-work chairs, made up all the
furniture in the room save for a square of Wilton carpet in the centre. The boards round and the
panelling of the walls were of brown, worm-eaten oak, so old and discoloured that it may have dated
from the original building of the house. Holmes drew one of the chairs into a corner and sat silent,
while his eyes travelled round and round and up and down, taking in every detail of the apartment.

"Where does that bell communicate with?" he asked at last pointing to a thick bell-rope which hung
down beside the bed, the tassel actually lying upon the pillow.

"It goes to the housekeeper's room."

"It looks newer than the other things?"

"Yes, it was only put there a couple of years ago."