184 Beaconside
"No?"

"Well, you know father didn't like anything of the sort. He wouldn't have any visitors if he could
help it, and he used to say that a woman should be happy in her own family circle. But then, as I
used to say to mother, a woman wants her own circle to begin with, and I had not got mine yet."