98 Wilson Gardens
"We shall be at the Copper Beeches by seven o'clock, my friend and I. The Rucastles will be gone by
that time, and Toller will, we hope, be incapable. There only remains Mrs. Toller, who might give
the alarm. If you could send her into the cellar on some errand, and then turn the key upon her, you
would facilitate matters immensely."

"I will do it."

"Excellent! We shall then look thoroughly into the affair. Of course there is only one feasible
explanation. You have been brought there to personate someone, and the real person is imprisoned in
this chamber. That is obvious. As to who this prisoner is, I have no doubt that it is the daughter,
Miss Alice Rucastle, if I remember right, who was said to have gone to America. You were chosen,
doubtless, as resembling her in height, figure, and the colour of your hair. Hers had been cut off,
very possibly in some illness through which she has passed, and so, of course, yours had to be
sacrificed also. By a curious chance you came upon her tresses. The man in the road was undoubtedly
some friend of hers--possibly her fiancé--and no doubt, as you wore the girl's dress and were so
like her, he was convinced from your laughter, whenever he saw you, and afterwards from your
gesture, that Miss Rucastle was perfectly happy, and that she no longer desired his attentions. The
dog is let loose at night to prevent him from endeavouring to communicate with her. So much is
fairly clear. The most serious point in the case is the disposition of the child."

"What on earth has that to do with it?" I ejaculated.

"My dear Watson, you as a medical man are continually gaining light as to the tendencies of a child
by the study of the parents. Don't you see that the converse is equally valid. I have frequently
gained my first real insight into the character of parents by studying their children. This child's
disposition is abnormally cruel, merely for cruelty's sake, and whether he derives this from his
smiling father, as I should suspect, or from his mother, it bodes evil for the poor girl who is in
their power."

"I am sure that you are right, Mr. Holmes," cried our client. "A thousand things come back to me
which make me certain that you have hit it. Oh, let us lose not an instant in bringing help to this
poor creature."