21 Holmefield Avenue
Lestrade shrugged his shoulders. "I am afraid that my colleague has been a little quick in forming
his conclusions," he said.

"But he is right. Oh! I know that he is right. James never did it. And about his quarrel with his
father, I am sure that the reason why he would not speak about it to the coroner was because I was
concerned in it."