29 Larch Walk
There was little difficulty in entering the grounds, for unrepaired breaches gaped in the old park
wall. Making our way among the trees, we reached the lawn, crossed it, and were about to enter
through the window when out from a clump of laurel bushes there darted what seemed to be a hideous
and distorted child, who threw itself upon the grass with writhing limbs and then ran swiftly across
the lawn into the darkness.

"My God!" I whispered; "did you see it?"