14 Ventnor Terrace
"I took the lamp from him, and I examined the machine very thoroughly. It was indeed a gigantic one,
and capable of exercising enormous pressure. When I passed outside, however, and pressed down the
levers which controlled it, I knew at once by the whishing sound that there was a slight leakage,
which allowed a regurgitation of water through one of the side cylinders. An examination showed that
one of the india-rubber bands which was round the head of a driving-rod had shrunk so as not quite
to fill the socket along which it worked. This was clearly the cause of the loss of power, and I
pointed it out to my companions, who followed my remarks very carefully and asked several practical
questions as to how they should proceed to set it right. When I had made it clear to them, I
returned to the main chamber of the machine and took a good look at it to satisfy my own curiosity.
It was obvious at a glance that the story of the fuller's-earth was the merest fabrication, for it
would be absurd to suppose that so powerful an engine could be designed for so inadequate a purpose.
The walls were of wood, but the floor consisted of a large iron trough, and when I came to examine
it I could see a crust of metallic deposit all over it. I had stooped and was scraping at this to
see exactly what it was when I heard a muttered exclamation in German and saw the cadaverous face of
the colonel looking down at me.

"'What are you doing there?' he asked.

"I felt angry at having been tricked by so elaborate a story as that which he had told me. 'I was
admiring your fuller's-earth,' said I; 'I think that I should be better able to advise you as to
your machine if I knew what the exact purpose was for which it was used.'

"The instant that I uttered the words I regretted the rashness of my speech. His face set hard, and
a baleful light sprang up in his grey eyes.