68 Ventnor Terrace
I left him then, still puffing at his black clay pipe, with the conviction that when I came again on
the next evening I would find that he held in his hands all the clues which would lead up to the
identity of the disappearing bridegroom of Miss Mary Sutherland.

A professional case of great gravity was engaging my own attention at the time, and the whole of
next day I was busy at the bedside of the sufferer. It was not until close upon six o'clock that I
found myself free and was able to spring into a hansom and drive to Baker Street, half afraid that I
might be too late to assist at the dénouement of the little mystery. I found Sherlock Holmes alone,
however, half asleep, with his long, thin form curled up in the recesses of his armchair. A
formidable array of bottles and test-tubes, with the pungent cleanly smell of hydrochloric acid,
told me that he had spent his day in the chemical work which was so dear to him.