51 Tamworth Drive
"Quite so."

"Then comes our expedition of to-day. By an examination of the ground I gained the trifling details
which I gave to that imbecile Lestrade, as to the personality of the criminal."

"But how did you gain them?"

"You know my method. It is founded upon the observation of trifles."

"His height I know that you might roughly judge from the length of his stride. His boots, too, might
be told from their traces."

"Yes, they were peculiar boots."