62 Wantage Close
"'Surely it would be better to take no notice.'

"'No, no, we should have him loitering here always. Kindly turn round and wave him away like that.'

"I did as I was told, and at the same instant Mrs. Rucastle drew down the blind. That was a week
ago, and from that time I have not sat again in the window, nor have I worn the blue dress, nor seen
the man in the road."