58 Tamworth Drive
"It is easy for me to be so, for every event of that dreadful time is seared into my memory. The
manor-house is, as I have already said, very old, and only one wing is now inhabited. The bedrooms
in this wing are on the ground floor, the sitting-rooms being in the central block of the buildings.
Of these bedrooms the first is Dr. Roylott's, the second my sister's, and the third my own. There is
no communication between them, but they all open out into the same corridor. Do I make myself
plain?"

"Perfectly so."

"The windows of the three rooms open out upon the lawn. That fatal night Dr. Roylott had gone to his
room early, though we knew that he had not retired to rest, for my sister was troubled by the smell
of the strong Indian cigars which it was his custom to smoke. She left her room, therefore, and came
into mine, where she sat for some time, chatting about her approaching wedding. At eleven o'clock
she rose to leave me, but she paused at the door and looked back.

"'Tell me, Helen,' said she, 'have you ever heard anyone whistle in the dead of the night?'