5 Harvey Street
"Why, it's a dummy," said he.

"Won't it ring?"

"No, it is not even attached to a wire. This is very interesting. You can see now that it is
fastened to a hook just above where the little opening for the ventilator is."

"How very absurd! I never noticed that before."