88 Valentia Road
"Good God!" I cried. "What can it mean, this relentless persecution?"

"The papers which Openshaw carried are obviously of vital importance to the person or persons in the
sailing-ship. I think that it is quite clear that there must be more than one of them. A single man
could not have carried out two deaths in such a way as to deceive a coroner's jury. There must have
been several in it, and they must have been men of resource and determination. Their papers they
mean to have, be the holder of them who it may. In this way you see K. K. K. ceases to be the
initials of an individual and becomes the badge of a society."

"But of what society?"

"Have you never--" said Sherlock Holmes, bending forward and sinking his voice--"have you never
heard of the Ku Klux Klan?"

"I never have."