106 Beaconside
"A certain selection and discretion must be used in producing a realistic effect," remarked Holmes.
"This is wanting in the police report, where more stress is laid, perhaps, upon the platitudes of
the magistrate than upon the details, which to an observer contain the vital essence of the whole
matter. Depend upon it, there is nothing so unnatural as the commonplace."

I smiled and shook my head. "I can quite understand your thinking so." I said. "Of course, in your
position of unofficial adviser and helper to everybody who is absolutely puzzled, throughout three
continents, you are brought in contact with all that is strange and bizarre. But here"--I picked up
the morning paper from the ground--"let us put it to a practical test. Here is the first heading
upon which I come. 'A husband's cruelty to his wife.' There is half a column of print, but I know
without reading it that it is all perfectly familiar to me. There is, of course, the other woman,
the drink, the push, the blow, the bruise, the sympathetic sister or landlady. The crudest of
writers could invent nothing more crude."